requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
google-generativeai
pandas
beautifulsoup4
